requirements.txt
imageio==2.28.1
numpy==1.24.3
Pillow==9.5.0
